% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimalistGODB.R
\name{grepList}
\alias{grepList}
\title{grepList}
\usage{
grepList(gaf)
}
\arguments{
\item{gaf}{character string containing the basename of the gaf file downloaded
from https://current.geneontology.org/products/pages/downloads.html}
}
\value{
returns the correct pattern to grep for
}
\description{
determine the correct pattern to grep for depending on the species
}
\examples{
pattern<-grepList("tair.gaf")

}
