% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimalistGODB.R
\name{restrictGOA}
\alias{restrictGOA}
\title{restrictGOA}
\usage{
restrictGOA(GOA, GO)
}
\arguments{
\item{GOA}{output of parseGOA()}

\item{GO}{output of parseGOBASIC()}
}
\value{
returns a restricted version of GOA
}
\description{
restrict GO categories in GOA to those in GO
}
\examples{
GOA<-restrictGOA(GOA,GO)

}
