% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimalistGODB.R
\name{subsetGOGOA}
\alias{subsetGOGOA}
\title{subsetGOGOA}
\usage{
subsetGOGOA(GOGOA)
}
\arguments{
\item{GOGOA}{return value of minimalistGODB::joinGO()}
}
\value{
returns a list containing subsets of GOGOA for each ontology, unique gene and cat lists, and stats
}
\description{
split GOGOA into 3 separate ontologies
}
\examples{
#load("data/GOGOAsmall.RData")
GOGOA3small<-subsetGOGOA(GOGOAsmall)

}
